💠 ID ile haber alma metodu eklendi

+++ app/src/main/java/com/iuce/news/db/dao/NewsWithStateDao.java
@@ -24,7 +24,7 @@ public interface NewsWithStateDao {
 
     @Transaction
     @Query("SELECT * FROM " + News.TABLE_NAME + " WHERE " + News.COLUMN_ID + " IN (:ids)")
-    LiveData<List<NewsWithState>> getNewsWithStateByIDs(Long... ids);
+    LiveData<List<NewsWithState>> getNewsWithStateByIDs(Integer... ids);
 
     @Transaction
     @Query("SELECT * FROM " + News.TABLE_NAME + " WHERE " + News.COLUMN_ID + " IN ("
